npc and enemy classes, and npc traveller
