+++ dnd-bot/dnd_bot/assets/maps/map.json
@@ -11,7 +11,7 @@
             [0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 27, 27, 0 , 0 , 0 , 2 , 2 , 0 , 2 , 0 , 56],
             [27, 27, 27, 27, 27, 27, 27, 22, 0 , 27, 27, 0 , 0 , 0 , 0 , 0 , 0 , 2 , 0 , 0 ],
             [27, 27, 27, 27, 27, 27, 21, 27, 0 , 27, 27, 0 , 0 , 0 , 0 , 0 , 0 , 0 , 2 , 2 ],
-            [1 , 0 , 1 , 0 , 1 , 0 , 22, 22, 0 , 27, 21, 27, 27, 27, 22, 0 , 0 , 54, 0 , 0 ],
+            [1 , 0 , 1 , 0 , 1 , 90, 22, 22, 0 , 27, 21, 27, 27, 27, 22, 0 , 0 , 54, 0 , 0 ],
             [0 , 1 , 0 , 1 , 0 , 0 , 0 , 0 , 0 , 22, 27, 27, 27, 27, 22, 0 , 0 , 0 , 4 , 0 ],
             [1 , 0 , 1 , 0 , 1 , 0 , 22, 22, 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 0 , 4 ],
             [27, 27, 27, 27, 27, 27, 21, 21, 27, 27, 27, 27, 27, 27, 22, 0 , 0 , 4 , 0 , 4 ]
@@ -50,6 +50,7 @@
         "53": "Lizardfolk archer",
         "54": "Nothic",
         "55": "Skeleton morningstar",
-        "56": "Skeleton warrior"
+        "56": "Skeleton warrior",
+        "90": "Traveller"
     }
 }
