+++ dnd-bot/dnd_bot/logic/prototype/entities/creatures/npc.py
@@ -1,0 +1,5 @@
+from dnd_bot.logic.prototype.creature import Creature
+
+
+class NPC(Creature):
+    pass
